requirements.txt
langchain
langchain-core
langchain-community
pymupdf
pypdf
tqdm
tqdm
sentence-transformers
faiss-cpu
chromadb
hf-xet
langchain-google-genai
python-dotenv
google-generativeai
generativeai
numpy